Fix default edx-themes user on edxapp
This fixes a problem with the default "themes_user" in
edxapp/meta/main.yml. The bug's main symptom would be failure to git
clone a theme configured via THEMES_REPOS.

Also, if the group is not set to common_web_group, failure to access the
directory by processes running under that group will cause server
errors.