App master as part of the multi-node stack
Allow launching the app master as part of the multi-node stack.  This
makes it more convenient to manage upgrades.

Another convenience change is exposing the app master OS::Nova:Server's
id as the physical_resource_id, instead of the stack's id.  Doing this
requires Kilo, however, and the template version was updated
accordingly.

Finally, update the README to use the `openstack` client, instead of
`heat`.