Add additional explicit dependencies in Heat template
Make the LBaaS pools explicitly depends on the load balancer. This
shouldn't be necessary, as there is already an indirect dependency via
the listener, but we've seen Heat misbehave on stack deletion, trying
to remove the load balancer before the pool is gone. Adding the
explicit dependencies ensures that upon deletion, the pools are
deleted before Heat deletes the load balancer.