Fix environment variables in suspender/reaper supervisor definitions
Juniper expects a different set of environment variables in supervisor
jobs than Ironwood did. Using the old set of variables means we're
missing LMS_CFG and REVISION_CFG, causing the suspender and reaper
supervisor jobs to fail on startup.

Use a modified form of the loop from the template that the edxapp role
uses to populate edxapp_env, in order to set the correct variables.