Update the hastexo_xblock role
Starting from v5 the Hastexo XBlock drops the Java servlet
application that has been used for `Guacamole` connections and uses
a Django ASGI application for the same instead.

Add conditions to check for the `hastexo_xblock_version` and either
run the tasks to deploy the Java application or use the
`hastexo_guacamole_client` role to deploy the Django application instead.